Fix onnx export of ivf_foward.onnx enablging moving coarse centroid search from javascript to onnx, change pq distance indices from int64 (default for pytorch) to int32 to potentially save memory

+++ ivfpq/python/src/ivfpq/pq.py
@@ -327,7 +327,7 @@ class PQDistance(nn.Module):
             total_distances, N, largest=False, sorted=True
         )
 
-        return indices, top_distances
+        return indices.to(torch.int32), top_distances
 
 
 def create_pq_system(
